bugfix: db_XXXX:{use} permission for arguments of user defined functions         make possiblity to leak unexpected information leak.
When a user defines a SQL function which inserts given arguments to others,
SE-PostgreSQL checks db_XXXX:{use} permission, but the function writes out
the given information via its arguments. In the result, user can read the
contents without db_XXXX:{select} permission.
This update changes the definition of {use} permission. It should be used
when we can assume the given arguments are not leaked to others.

The current implementation assumes built-in functions are not malicious,
but user-defined functions are not.